requirements.txt
torch>=2.0.0
transformers>=4.30.0
Pillow>=9.0.0
TTS>=0.13.0
groq>=0.3.0
python-dotenv>=1.0.0
accelerate>=0.20.0
SpeechRecognition
PyAudio
gtts